+++ CHANGELOG.md
@@ -1,9 +1,11 @@
 # Changelog
 
-## Next Release
+## Next release
 
 ### New features
 
+- Added `Wizaplace\SDK\Basket\Payment::getExternalReference`
+- Added `Wizaplace\SDK\Order\Payment::getExternalReference`
 - Added `Wizaplace\SDK\Catalog::getProductsByMvpId`
 - Added parameter `bool $allowMvp = true` to `Wizaplace\SDK\Catalog::getProductsBy` methods